restructuring the code
