+++ lightfsm.wpfexmpl/Classes/ViewsFsm/States/CheckVersionStateImpl.cs
@@ -1,0 +1,10 @@
+﻿namespace LightStateMachine.WpfExample.Classes
+{
+    using System.Collections.Generic;
+    using System.Threading.Tasks;
+    using LightStateMachine.Interfaces;
+
+    public class CheckVersionStateImpl : StateImplBase
+    {
+    }
+}
