+++ lightfsm.wpfexmpl/Classes/ViewsFsm/States/ExitStateImpl.cs
@@ -1,0 +1,9 @@
+﻿namespace LightStateMachine.WpfExample.Classes
+{
+    using System.Collections.Generic;
+    using System.Threading.Tasks;
+    using LightStateMachine.Interfaces;
+    public class ExitStateImpl : StateImplBase
+    {
+    }
+}
