+++ lightfsm.wpfexmpl/Classes/ViewsFsm/ViewsStateConfiguration.cs
@@ -1,39 +1,7 @@
 ﻿namespace LightStateMachine.WpfExample.Classes
 {
     using System.Collections.Generic;
-    using System.Threading.Tasks;
     using LightStateMachine.Interfaces;
-
-    /// <summary>
-    /// The Backup files states enum.
-    /// </summary>
-    public enum BackupFilesStateEnum
-    {
-        /// <summary>
-        /// The check version state.
-        /// </summary>
-        SettingsState,
-
-        /// <summary>
-        /// Transfer to removeable state.
-        /// </summary>
-        TransferToRemoveableState,
-
-        /// <summary>
-        /// The recover previous version state.
-        /// </summary>
-        TransferToLocationState,
-
-        /// <summary>
-        /// The exit state.
-        /// </summary>
-        ExitState,
-
-        /// <summary>
-        /// The null state.
-        /// </summary>
-        NullState
-    }
 
     /// <summary>
     /// The states configurator.
@@ -48,12 +16,12 @@
         }
 
         /// <summary>
-        /// The exit state enum.
+        /// Gets the exit state enum.
         /// </summary>
         public BackupFilesStateEnum StartState => BackupFilesStateEnum.SettingsState;
 
         /// <summary>
-        /// The exit state enum.
+        /// Gets the exit state enum.
         /// </summary>
         public BackupFilesStateEnum ExitState => BackupFilesStateEnum.ExitState;
 
@@ -73,38 +41,5 @@
                            { BackupFilesStateEnum.ExitState, new ExitStateImpl() },
                        };
         }
-    }
-
-    public class StateImplBase : IStateAction<BackupFilesStateEnum, object>
-    {
-        public bool CanPerform()
-        {
-            throw new System.NotImplementedException();
-        }
-
-        public Task<BackupFilesStateEnum> PerformTransitionAction(object payload)
-        {
-            throw new System.NotImplementedException();
-        }
-
-        public void PerformOnEnterAction(object payload)
-        {
-            throw new System.NotImplementedException();
-        }
-    }
-    public class ExitStateImpl : StateImplBase
-    {
-    }
-
-    public class PreviousVersionRecoveryStateImpl : StateImplBase
-    {
-    }
-
-    public class InstallStateImpl : StateImplBase
-    {
-    }
-
-    public class CheckVersionStateImpl : StateImplBase
-    {
     }
 }
