+++ lightfsm.wpfexmpl/lightfsm.wpfexmpl.csproj
@@ -69,6 +69,12 @@
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </ApplicationDefinition>
+    <Compile Include="Classes\ViewsFsm\BackupFilesStateEnum.cs" />
+    <Compile Include="Classes\ViewsFsm\States\CheckVersionStateImpl.cs" />
+    <Compile Include="Classes\ViewsFsm\States\ExitStateImpl.cs" />
+    <Compile Include="Classes\ViewsFsm\States\InstallStateImpl.cs" />
+    <Compile Include="Classes\ViewsFsm\States\PreviousVersionRecoveryStateImpl.cs" />
+    <Compile Include="Classes\ViewsFsm\States\StateImplBase.cs" />
     <Compile Include="DelegateCommand.cs" />
     <Compile Include="ViewModelBase.cs" />
     <Page Include="MainWindow.xaml">
@@ -79,7 +85,7 @@
       <DependentUpon>App.xaml</DependentUpon>
       <SubType>Code</SubType>
     </Compile>
-    <Compile Include="Classes\ViewsStateConfiguration.cs" />
+    <Compile Include="Classes\ViewsFsm\ViewsStateConfiguration.cs" />
     <Compile Include="MainWindow.xaml.cs">
       <DependentUpon>MainWindow.xaml</DependentUpon>
       <SubType>Code</SubType>
@@ -124,5 +130,6 @@
       <Name>core.lightfsm</Name>
     </ProjectReference>
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
